% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_goi}
\alias{filter_goi}
\title{Filter a list of predictor data frames by genes of interest}
\usage{
filter_goi(pred_list, goi)
}
\arguments{
\item{pred_list}{A list of tibbles or data frames, each containing a column named \code{gene}.}

\item{goi}{A character vector of genes of interest to retain.}
}
\value{
A list of filtered tibbles with only the genes of interest.
}
\description{
This function takes a list of tibbles (each containing a \code{gene} column)
and filters each one to include only the specified genes of interest.
}
\examples{
\dontrun{
filtered_list <- filter_goi(pred_list, c("GeneA", "GeneB", "GeneC"))
}

}
